+++ src/actions.js
@@ -1,9 +1,3 @@
-import {
-  fetchCategories,
-  fetchRegions,
-  fetchRestaurants,
-} from './services/api';
-
 export function setRegions(regions) {
   return {
     type: 'setRegions',
@@ -37,26 +31,5 @@ export function setSelected(selected) {
     payload: {
       selected,
     },
-  };
-}
-
-export function loadRegions() {
-  return async (dispatch) => {
-    const regions = await fetchRegions();
-    dispatch(setRegions(regions));
-  };
-}
-
-export function loadCategories() {
-  return async (dispatch) => {
-    const categories = await fetchCategories();
-    dispatch(setCategories(categories));
-  };
-}
-
-export function loadRestaurnats({ region, category }) {
-  return async (dispatch) => {
-    const restaurants = await fetchRestaurants({ region, category });
-    dispatch(setRestaurnats(restaurants));
   };
 }
