+++ src/components/RegionsContainer.jsx
@@ -2,13 +2,13 @@ import React from 'react';
 
 import { useSelector, useDispatch } from 'react-redux';
 
-import { setSelected } from './actions';
+import { setSelected } from '../actions';
 
-import Buttons from './Buttons';
+import Buttons from '../views/Buttons';
 
-export default function CategoriesContainer() {
-  const { categories, selected: { category, region } } = useSelector((state) => ({
-    categories: state.categories,
+export default function RegionsContainer() {
+  const { regions, selected: { category, region } } = useSelector((state) => ({
+    regions: state.regions,
     selected: state.selected,
   }));
 
@@ -17,22 +17,22 @@ export default function CategoriesContainer() {
   const handleClick = (event) => {
     const { name, id } = event.target;
 
-    if (!(name === category.name)) {
+    if (!(name === region.name)) {
       dispatch(setSelected({
-        region,
-        category: {
+        region: {
           id,
           name,
         },
+        category,
       }));
     }
   };
 
   return (
     <Buttons
-      buttons={categories}
+      buttons={regions}
       handleClick={handleClick}
-      selectedName={category.name}
+      selectedName={region.name}
     />
   );
 }
