+++ src/components/index.js
@@ -1,0 +1,3 @@
+export { default as RegionsContainer } from './RegionsContainer';
+export { default as CategoriesContainer } from './CategoriesContainer';
+export { default as RestaurantsContainer } from './RestaurantsContainer';
