+++ src/index.jsx
@@ -6,7 +6,7 @@ import { Provider } from 'react-redux';
 
 import store from './store';
 
-import App from './App';
+import App from './views/App';
 
 ReactDOM.render(
   <Provider store={store}>
