+++ src/views/App.test.jsx
@@ -6,7 +6,7 @@ import { useSelector, useDispatch } from 'react-redux';
 
 import given from 'given2';
 
-import { restaurants } from '../fixtures';
+import { restaurants } from '../../fixtures';
 
 import App from './App';
