+++ src/views/Restaurants.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { render } from '@testing-library/react';
 
-import { restaurants } from '../fixtures';
+import { restaurants } from '../../fixtures';
 
 import Restaurants from './Restaurants';
